+++ code/src/cljs/sixsq/nuvla/ui/edges_detail/views.cljs
@@ -233,7 +233,7 @@
                  :header  (@tr [:nuvlabox-update-warning])
                  :content (r/as-element
                             [:span (str (@tr [:nuvlabox-update-error-content])) " "
-                             [:a {:href   "https://docs.nuvla.io/nuvlabox/nuvlabox-engine/v2/installation/"
+                             [:a {:href   "https://docs.nuvla.io/nuvlaedge/installation/"
                                   :target "_blank"}
                               (str/capitalize (@tr [:see-more]))]])}])
              (when (and (some? target-version) (is-old-version? target-version))