+++ CHANGELOG.md
@@ -2,6 +2,10 @@
 
 ## Unreleased
 
+### Changed
+
+- Profile - Groups tabs warns user of unsaved changes
+
 ## [2.32.5] - 2022-09-19
 
 - Home - Update old links to point to new NuvlaEdge documentation