+++ CHANGELOG.md
@@ -4,6 +4,7 @@
 
 ### Changed
 
+- Home - Update old links to point to new NuvlaEdge documentation
 - Profile - Groups tabs warns user of unsaved changes
 
 ## [2.32.5] - 2022-09-19